+++ app/Models/Age.php
@@ -10,4 +10,8 @@ class Age extends Model
     public $timestamps = false;
 
     use HasFactory;
+    public function respondents()
+    {
+        return $this->hasMany(Respondent::class);
+    }
 }