+++ app/Models/Category.php
@@ -10,4 +10,14 @@ class Category extends Model
     public $timestamps = false;
 
     use HasFactory;
+
+    /**
+     * Get all of the respondents for the Category
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\HasMany
+     */
+    public function respondents()
+    {
+        return $this->hasMany(Respondent::class);
+    }
 }