+++ app/Models/City.php
@@ -10,4 +10,13 @@ class City extends Model
     public $timestamps = false;
 
     use HasFactory;
+    /**
+     * Get all of the respondents for the City
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\HasMany
+     */
+    public function respondents()
+    {
+        return $this->hasMany(Respondent::class);
+    }
 }