+++ resources/views/app/admin.blade.php
@@ -1,0 +1,7 @@
+@extends('layout.survey')
+@section('content')
+<div id="app">
+  {{-- <example-component/> --}}
+</div>
+<script src="{{ mix('/js/admin/app.js') }}"></script>
+@endsection